protocol: add helpers to convert cursor to/from objects
**Summary**

Sometimes users need to store the cursor for later use (e.g. to restart
from where the indexer left off). The `Cursor` object from gRPC is
clunky to work with, so we provide two helper functions to convert it to
and from plain Javascript objects.

**Test Plan**

We test that the cursor can be converted back and forth between the gRPC
representation and the Javascript one.